fix: resolve all CI gate failures — UUID session round-trips, ORG016 violations, block_on panics, and config env vars
- Fix session create handler to emit valid UUIDs (not prefixed strings) so
  SessionId::from_string round-trips correctly instead of falling back to v5
- Fix empty session_id validation in agent handler for UUID-based IDs
  (SessionId::from_name("") produces a valid v5 UUID, not empty string)
- Replace tokio block_on with dedicated thread+runtime in declarative_validator
  to avoid panic when called inside an existing async runtime
- Fix 5 ORG016 errors: remove dead FileTreeBuilder (zero callers) from domain
  test_utils.rs, inline OrgContext::current() → OrgContext::default() at call site
- Fix corrupt DB recovery in SQLite provider to handle pragma failures (not just DDL)
- Fix env var MCP__AUTH__USER_DB_PATH → MCP__PROVIDERS__DATABASE__CONFIGS__DEFAULT__PATH
  in startup smoke and stdio transport integration tests
- Remove #[serde(default)] from ModeConfig (config defaults come from default.toml)
- Update hooks integration test assertions to use dynamic UUID values
- Centralize constants (CONTENT_TYPE_JSON, language detection, search defaults)
  into infrastructure/providers crates with proper re-exports

All three CI gates now pass:
- make lint MCB_CI=1 → exit 0
- make test THREADS=4 → 1671 passed, 0 failed
- make validate → exit 0 (zero errors)

+++ crates/mcb-providers/tests/unit/hybrid_search_tests.rs
@@ -2,10 +2,10 @@
 
 #![cfg(feature = "hybrid-search")]
 
+use mcb_domain::constants::search::{HYBRID_SEARCH_BM25_WEIGHT, HYBRID_SEARCH_SEMANTIC_WEIGHT};
 use mcb_domain::entities::CodeChunk;
 use mcb_domain::ports::providers::HybridSearchProvider;
 use mcb_domain::value_objects::SearchResult;
-use mcb_providers::constants::{HYBRID_SEARCH_BM25_WEIGHT, HYBRID_SEARCH_SEMANTIC_WEIGHT};
 use mcb_providers::hybrid_search::{BM25Params, BM25Scorer, HybridSearchEngine};
 use rstest::rstest;
 
@@ -138,9 +138,9 @@ fn bm25_relevant_chunk_ranks_higher(#[case] mode: &str) {
 async fn hybrid_search_engine_creation(#[case] weight_kind: &str) {
     let engine = HybridSearchEngine::new();
     if weight_kind == "bm25" {
-        assert!((engine.bm25_weight() - HYBRID_SEARCH_BM25_WEIGHT).abs() < f32::EPSILON);
+        assert!((engine.bm25_weight() - HYBRID_SEARCH_BM25_WEIGHT).abs() < f64::EPSILON);
     } else {
-        assert!((engine.semantic_weight() - HYBRID_SEARCH_SEMANTIC_WEIGHT).abs() < f32::EPSILON);
+        assert!((engine.semantic_weight() - HYBRID_SEARCH_SEMANTIC_WEIGHT).abs() < f64::EPSILON);
     }
 }
